rename: generate domain package
