+++ src/main/java/com/walletguardians/walletguardiansapi/domain/category/controller/CategoryController.java
@@ -1,15 +1,12 @@
-package com.walletguardians.walletguardiansapi.category.controller;
+package com.walletguardians.walletguardiansapi.domain.category.controller;
 
-import com.walletguardians.walletguardiansapi.category.controller.dto.response.CategoryResponse;
-import com.walletguardians.walletguardiansapi.category.entity.Category;
-import com.walletguardians.walletguardiansapi.category.service.CategoryService;
+import com.walletguardians.walletguardiansapi.domain.category.controller.dto.response.CategoryResponse;
+import com.walletguardians.walletguardiansapi.domain.category.service.CategoryService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @RestController
