+++ src/main/java/com/walletguardians/walletguardiansapi/domain/category/repository/CategoryRepository.java
@@ -1,0 +1,7 @@
+package com.walletguardians.walletguardiansapi.domain.category.repository;
+
+import com.walletguardians.walletguardiansapi.domain.category.entity.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+}
